ConsoleMenuView updated
